HACK week: WIP, payments details
