+++ client/payment-details/summary/test/index.tsx
@@ -1,0 +1,146 @@
+/** @format */
+/**
+ * External dependencies
+ */
+import { render } from '@testing-library/react';
+import React from 'react';
+
+/**
+ * Internal dependencies
+ */
+import PaymentDetailsSummary from '../';
+import { Charge } from 'wcpay/types/charges';
+
+declare const global: {
+	wcpaySettings: {
+		isSubscriptionsActive: boolean;
+		zeroDecimalCurrencies: string[];
+	};
+};
+
+const getBaseCharge = (): Charge =>
+	( {
+		id: 'ch_38jdHA39KKA',
+		/* Stripe data comes in seconds, instead of the default Date milliseconds */
+		created: Date.parse( 'Sep 19, 2019, 5:24 pm' ) / 1000,
+		amount: 2000,
+		amount_refunded: 0,
+		application_fee_amount: 70,
+		disputed: false,
+		dispute: null,
+		currency: 'usd',
+		type: 'charge',
+		status: 'succeeded',
+		paid: true,
+		captured: true,
+		balance_transaction: {
+			amount: 2000,
+			currency: 'usd',
+			fee: 70,
+		},
+		refunds: {
+			data: [],
+		},
+		order: {
+			number: 45981,
+			url: 'https://somerandomorderurl.com/?edit_order=45981',
+		},
+		billing_details: {
+			name: 'Customer name',
+		},
+		payment_method_details: {
+			card: {
+				brand: 'visa',
+				last4: '4242',
+			},
+			type: 'card',
+		},
+		outcome: {
+			risk_level: 'normal',
+		},
+	} as any );
+
+function renderCharge( charge: Charge, isLoading = false ) {
+	const { container } = render(
+		<PaymentDetailsSummary charge={ charge } isLoading={ isLoading } />
+	);
+	return container;
+}
+
+describe( 'PaymentDetailsSummary', () => {
+	beforeEach( () => {
+		global.wcpaySettings = {
+			isSubscriptionsActive: false,
+			zeroDecimalCurrencies: [],
+		};
+	} );
+
+	test( 'correctly renders a charge', () => {
+		expect( renderCharge( getBaseCharge() ) ).toMatchSnapshot();
+	} );
+
+	test( 'renders partially refunded information for a charge', () => {
+		const charge = getBaseCharge();
+		charge.refunded = false;
+		charge.amount_refunded = 1200;
+		charge.refunds.data.push( {
+			balance_transaction: {
+				amount: -charge.amount_refunded,
+				currency: 'usd',
+			} as any,
+		} );
+
+		expect( renderCharge( charge ) ).toMatchSnapshot();
+	} );
+
+	test( 'renders fully refunded information for a charge', () => {
+		const charge = getBaseCharge();
+		charge.refunded = true;
+		charge.amount_refunded = 2000;
+		charge.refunds.data.push( {
+			balance_transaction: {
+				amount: -charge.amount_refunded,
+				currency: 'usd',
+			} as any,
+		} );
+
+		expect( renderCharge( charge ) ).toMatchSnapshot();
+	} );
+
+	test( 'renders the information of a disputed charge', () => {
+		const charge = getBaseCharge();
+		charge.disputed = true;
+		charge.dispute = {
+			amount: 1500,
+			status: 'under_review',
+			balance_transactions: [
+				{
+					amount: -1500,
+					fee: 1500,
+				} as any,
+			],
+		} as any;
+
+		expect( renderCharge( charge ) ).toMatchSnapshot();
+	} );
+
+	test( 'renders a charge with subscriptions', () => {
+		global.wcpaySettings.isSubscriptionsActive = true;
+
+		const charge = getBaseCharge();
+		if ( charge.order ) {
+			charge.order.subscriptions = [
+				{
+					number: 246,
+					url: 'https://example.com/subscription/246',
+				},
+			];
+		}
+
+		expect( renderCharge( charge ) ).toMatchSnapshot();
+	} );
+
+	test( 'renders loading state', () => {
+		expect( renderCharge( {} as any, true ) ).toMatchSnapshot();
+	} );
+} );
